api/requirements.txt
fastapi==0.115.6
python-multipart==0.0.20
python-dotenv==1.0.1
nest-asyncio==1.6.0
llama-index-core==0.14.9
llama-index-llms-google-genai==0.1.4
llama-index-embeddings-google-genai==0.1.3
llama-cloud-services==0.6.54
